refactor tests: centralize app rendering
