+++ apps/react/tests/music-notation-rest-test.tsx
@@ -1,17 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { Provider } from 'react-redux';
 import { MusicNotation } from '../src/components/MusicNotation';
 import '../src/index.css';
-import { createStore } from 'MemoryFlashCore/src/redux/store';
 import { MultiSheetQuestion } from 'MemoryFlashCore/src/types/MultiSheetCard';
 import { StaffEnum } from 'MemoryFlashCore/src/types/Cards';
-
-const indexParam = new URLSearchParams(window.location.search).get('index');
-const multiPartCardIndex = indexParam ? parseInt(indexParam, 10) : 0;
-
-const store = createStore({ scheduler: { multiPartCardIndex } } as any, () => {});
-(window as any).store = store;
+import { renderApp } from './renderApp';
 
 const data: MultiSheetQuestion = {
 	key: 'C',
@@ -28,10 +20,4 @@ const data: MultiSheetQuestion = {
 	],
 };
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
-	<React.StrictMode>
-		<Provider store={store}>
-			<MusicNotation data={data} highlightClassName="highlight" />
-		</Provider>
-	</React.StrictMode>,
-);
+renderApp(<MusicNotation data={data} highlightClassName="highlight" />);
