+++ apps/react/tests/music-recorder-chord-test.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { Provider } from 'react-redux';
 import { MusicNotation } from '../src/components/MusicNotation';
 import '../src/index.css';
-import { createStore } from 'MemoryFlashCore/src/redux/store';
 import { MusicRecorder } from 'MemoryFlashCore/src/lib/MusicRecorder';
-
-const store = createStore({ scheduler: { multiPartCardIndex: 0 } } as any, () => {});
-(window as any).store = store;
+import { renderApp } from './renderApp';
 
 const recorder = new MusicRecorder('q');
 (window as any).recorder = recorder;
@@ -22,10 +17,4 @@ const App: React.FC = () => {
 	return <MusicNotation data={data} />;
 };
 
-ReactDOM.createRoot(document.getElementById('root')!).render(
-	<React.StrictMode>
-		<Provider store={store}>
-			<App />
-		</Provider>
-	</React.StrictMode>,
-);
+renderApp(<App />);
